Evaluate_Source/requirements.txt
uvicorn
fastAPI
FlagEmbedding
fastapi
cmake
sqlalchemy
selenium
numpy==1.26.4
uvicorn
pydantic
typing
psycopg2-binary
python-docx
gdown
selenium
transformers==4.52.3 
pandas==2.2.3 
openpyxl 
xlsxwriter
deep-translator
langid
wikipedia
gguf==0.16.3
bitsandbytes==0.46.0 
accelerate==1.7.0 
peft==0.15.2
llama_cpp_python==0.3.9
openai
psycopg2-binary
google-genai
google-generativeai
httpx==0.28.1
googletrans
googlesearch-python
sentence_transformers
torch
docx
FlagEmbedding
jupyter
papermill
youtube_transcript_api
nlkt